requirements.txt
beautifulsoup4
typer-slim